imp: unified playerId
Instead of having a player entity id, a network id, and a user id, use a single player id for all of these.

This makes it much simpler for app scripts to manage everything.

This also introduces the limitation that a single user can only exist in the world once, meaning if you are testing multiplayer logic you will need to open a new window in incognito to generate a new user. Ultimately this is a good trade-off for a better app developer experience.