feat: ui scaler option
When creating UI in world-space you sometimes want it to scale as if it was in screen-space, so that when you are far away it scales up to match the size it would be if it were on screen and vice versa.

This is useful for things like buttons and chat bubbles.